requirements.txt
accelerate==1.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.13
aioprometheus==23.12.0
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
async-timeout==5.0.1
attrs==25.1.0
bitsandbytes==0.45.5
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cmake==3.31.6
colorama==0.4.6
contexttimer==0.3.3
datasets==3.4.0
dill==0.3.8
distro==1.9.0
exceptiongroup==1.2.2
fastapi==0.115.11
filelock==3.17.0
fire==0.7.0
Flask==3.1.0
frozenlist==1.5.0
fsspec==2024.12.0
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.29.3
idna==3.10
importlib_metadata==8.6.1
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.9.0
joblib==1.4.2
json_repair==0.44.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
lit==18.1.8
loguru==0.7.3
MarkupSafe==3.0.2
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
multiprocess==0.70.16
networkx==3.2.1
ninja==1.11.1.3
numpy==2.0.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu11==10.2.10.91
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu11==11.7.4.91
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu11==2.14.3
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu11==11.7.91
nvidia-nvtx-cu12==12.1.105
openai==1.66.3
orjson==3.10.15
packaging==24.2
pandas==2.2.3
peft==0.13.2
propcache==0.3.0
protobuf==6.30.0
psutil==7.0.0
pyarrow==19.0.1
pybind11==2.13.6
pydantic==2.10.6
pydantic_core==2.27.2
pynvml==11.5.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
quantile-python==1.1
rank-bm25==0.2.2
ray==2.43.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rpds-py==0.23.1
sacremoses==0.1.1
safetensors==0.5.3
sentencepiece==0.2.0
six==1.17.0
sniffio==1.3.1
starlette==0.46.1
sympy==1.13.3
tenacity==9.1.2
termcolor==3.1.0
tokenizers==0.15.2
torch==2.1.2
tqdm==4.67.1
transformers==4.36.0
transformers-stream-generator==0.0.5
tree-sitter==0.23.2
tree-sitter-python==0.23.6
triton==2.1.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0
vllm==0.3.0
watchfiles==1.0.4
websockets==15.0.1
Werkzeug==3.1.3
xformers==0.0.23.post1
xxhash==3.5.0
yarl==1.18.3
zipp==3.21.0
